bit_util.c attempted snapToSprite

+++ include/collision/bit_util.h
@@ -4,6 +4,7 @@
 #include "common.h"
 #include "common/obj.h"
 #include "common/level.h"
+#include "collision/block_22C84.h"
 
 /* TODO: is this related with collision at all? */
 
@@ -14,7 +15,7 @@ void set2bits(s32 *param_1, u8 pos, s32 param_3);
 void read2bits(s32 *value, u8 pos, s32 *b1, s32 *b0);
 void save_objects_flags(void);
 void restore_objects_flags(void);
-void snapToSprite(Obj *obj1, Obj *obj2, u32 param_3, s32 param_4, u16 param_5);
+void snapToSprite(Obj *obj1, Obj *obj2, u8 param_3, s32 param_4, u16 param_5);
 Obj * findfirstObject(ObjType obj_type);
 Obj * findfirstInactiveObject(ObjType param_1);
 void setbit(u8 *buffer, u32 bit_index);
